+++ Classes/Controller/Example/BeUsersController.php
@@ -16,15 +16,15 @@ class BeUsersController extends AbstractBackendController
         return ['be_users', 'be_groups', 'sys_filemounts'];
     }
 
+    public function getTemplateConfigurations(): array
+    {
+        return ['Example/BeUsers' => []];
+    }
+
     protected function modifyTableConfiguration(): void
     {
         $this->tableConfiguration['be_users']['columns']['username']['defaultPosition'] = 1;
         $this->tableConfiguration['be_users']['columns']['realName']['defaultPosition'] = 2;
         $this->tableConfiguration['be_users']['columns']['email']['defaultPosition'] = 3;
-    }
-
-    protected function getTemplateName(): string
-    {
-        return 'Example/BeUsers';
     }
 }